requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
google-generativeai==0.3.2
pydantic==2.5.3
python-dotenv==1.0.0
sqlalchemy==2.0.25
httpx==0.26.0
groq==0.4.1
